% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clasificarContactosApp.R
\name{clasificarContactosApp}
\alias{clasificarContactosApp}
\title{clasificarContactosApp}
\usage{
clasificarContactosApp(path)
}
\arguments{
\item{path, }{string}
}
\description{
funcion principal del paquete de clasificarContactos
}
\author{
Borja Cobos
}
