% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generarModelo.R
\name{seleccionColumnas}
\alias{seleccionColumnas}
\title{seleccionarColumnas}
\usage{
seleccionColumnas(datosSplit, config)
}
\arguments{
\item{datosSplit}{}

\item{config}{}
}
\description{
seleccionarColumnas
}
